+++ src/test/java/wooteco/subway/admin/acceptance/LineStationAcceptanceTest.java
@@ -2,6 +2,8 @@ package wooteco.subway.admin.acceptance;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.LineResponse;
@@ -32,5 +34,25 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
 
         LineDetailResponse lineResponseAfterRemoveLineStation = getLine(lineResponse.getId());
         assertThat(lineResponseAfterRemoveLineStation.getStations().size()).isEqualTo(2);
+    }
+
+    LineDetailResponse getLine(Long id) {
+        return
+                given().when().
+                        get("/lines/" + id).
+                        then().
+                        log().all().
+                        extract().as(LineDetailResponse.class);
+    }
+
+    void removeLineStation(Long lineId, Long stationId) {
+        given().
+                contentType(MediaType.APPLICATION_JSON_VALUE).
+                accept(MediaType.APPLICATION_JSON_VALUE).
+                when().
+                delete("/lines/" + lineId + "/stations/" + stationId).
+                then().
+                log().all().
+                statusCode(HttpStatus.NO_CONTENT.value());
     }
 }
