+++ src/test/java/wooteco/subway/admin/service/PathServiceTest.java
@@ -1,7 +1,6 @@
 package wooteco.subway.admin.service;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -105,19 +104,5 @@ class PathServiceTest {
         assertThatThrownBy(() -> pathService.findPath(shortestPathResponse))
                 .isInstanceOf(PathException.class)
                 .hasMessage("해당 역을 찾을 수 없습니다.");
-    }
-
-    @DisplayName("한글이외의 값이 입력될경우 익셉션이 발생한다")
-    @Test
-    void wrongInputExceptionTest() {
-        String source = "gangNamyeok";
-        String target = "잠실역";
-        String pathType = "DISTANCE";
-
-        ShortestPathResponse shortestPathResponse = new ShortestPathResponse(source, target, pathType);
-
-        assertThatThrownBy(() -> pathService.findPath(shortestPathResponse))
-                .isInstanceOf(PathException.class)
-                .hasMessage("출발역과 도착역은 한글만 입력가능합니다.");
     }
 }
